5xx Support
Alright everyone, we now have full support for all 5.xx! We're still working on 6.50 and 7.61, which will be available later. Give it a try and let us know your feedback.

A huge thanks to @sleirsgoevy for creating this project! Also, a big thanks to @buzzer-re , @idlesauce , @BestPig  ,@LightningMods , @zecoxao  and his keys friend, and everyone else who contributed. We truly appreciate all the help!